fast_api/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
black==24.8.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dill==0.3.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
peewee==3.17.6
platformdirs==4.3.2
pydantic==2.9.1
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.23.3
Pygments==2.18.0
pylint==3.2.7
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
tomlkit==0.13.2
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
